python/requirements.txt
cryptography==42.0.8
PyJWT==2.8.0
